Revert refactor into SpaceProposalsContent and fix loading
